+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Changelog
 
 ## v0.8.4-dev
+* Enhancements
+  * support rename parameter using `source`
 
 ## v0.8.3 (2015-10-03)
 * Enhancements